moved timer to sources
